# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SkillUp Platform API
  description: |
    SkillUp Platform 提供全面的在线学习平台 API，包括用户认证、课程管理、
    人脸识别、监控统计等功能。
    
    ## 特性
    - JWT 认证
    - 人脸识别和验证
    - 短信验证
    - 学习进度跟踪
    - 监控统计
    - 错误追踪
    
    ## 环境
    - 开发环境: https://dev-api.skillup.com
    - 测试环境: https://staging-api.skillup.com
    - 生产环境: https://api.skillup.com
  version: 1.0.0
  contact:
    name: SkillUp Platform API Support
    email: api-support@skillup.com
    url: https://docs.skillup.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.skillup.com/v1
    description: 生产环境
  - url: https://staging-api.skillup.com/v1
    description: 测试环境
  - url: https://dev-api.skillup.com/v1
    description: 开发环境
security:
  - BearerAuth: []
paths:
...
  /api/monitoring/stats:
    get:
      tags:
        - Monitoring
      summary: 获取监控统计数据
      description: 获取系统监控统计信息，支持多种导出格式
      parameters:
        - name: startDate
          in: query
          description: 开始日期
          schema:
            type: string
            format: date
            example: "2025-08-01"
        - name: endDate
          in: query
          description: 结束日期
          schema:
            type: string
            format: date
            example: "2025-08-31"
        - name: service
          in: query
          description: 服务名称
          schema:
            type: string
            example: "skillup-platform"
        - name: environment
          in: query
          description: 环境
          schema:
            type: string
            enum: [development, staging, production]
            example: "production"
        - name: format
          in: query
          description: 导出格式
          schema:
            type: string
            enum: [json, csv, xlsx]
            default: json
        - name: groupBy
          in: query
          description: 分组方式
          schema:
            type: string
            enum: [hour, day, week]
            default: day
      responses:
        '200':
          description: 统计数据获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringStatsResponse'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: false
            error:
              type: object
              properties:
                code:
                  type: string
                  example: "VALIDATION_ERROR"
                message:
                  type: string
                  example: "参数验证失败"
                details:
                  type: string
                requestId:
                  type: string
                  example: "req_12345"
...
    MonitoringStatsResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                stats:
                  type: array
                  items:
                    type: object
                    properties:
                      date:
                        type: string
                        format: date
                      hour:
                        type: integer
                      service:
                        type: string
                      environment:
                        type: string
                      totalRequests:
                        type: integer
                      errorRequests:
                        type: integer
                      avgResponseTime:
                        type: number
                      p95ResponseTime:
                        type: number
                      p99ResponseTime:
                        type: number
                metadata:
                  type: object
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未授权访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...